fixed openmp on windows

+++ client/src/world.cpp
@@ -3,7 +3,6 @@
 #include "../include/world.hpp"
 
 #include <iostream>
-
 #include <omp.h>
 
 World::World(Settings st, SDL_Renderer *renderer) : size(st.world_generation.resolution) {
@@ -22,9 +21,9 @@ World::World(Settings st, SDL_Renderer *renderer) : size(st.world_generation.res
 
     pixels = static_cast<uint32_t *>(surf->pixels);
 
-    #pragma omp parallel for collapse(2)
-    for (size_t y = 0; y < height; y++) {
-        for (size_t x = 0; x < width; x++) {
+#pragma omp parallel for collapse(2)
+    for (int y = 0; y < height; y++) {
+        for (int x = 0; x < width; x++) {
             // get the terrain generation data form the settings file
             float freq      = 7.68f * st.world_generation.frequency / width;
             float height    = 0.0f;
